Switch the site over to sidebar layout
