+++ scss/_c-hero.scss
@@ -12,7 +12,7 @@
   flex-direction: column-reverse;
 
   @media (min-width: $g-bp-2) {
-    padding: 9.6rem 0;
+    padding: 0 0 9.6rem;
     flex-direction: row;
     justify-content: space-between;
     align-items: center;
