+++ scss/_c-navigation-mobile.scss
@@ -43,6 +43,7 @@
   color: #1D2C48;
   display: flex;
   padding: 1.6rem;
+  align-items: center;
   justify-content: space-between;
   svg {
     opacity: 0.5;
