+++ scss/_e-base.scss
@@ -22,7 +22,6 @@ html {
 }
 
 body {
-  border-top: 0.4rem solid #E938C2;
   background: #FFF;
   color: $g-text-color;
   font-family: $g-text-font-family;
