+++ src/components/Sidebar.svelte
@@ -1,0 +1,11 @@
+<script>
+    import NavItems from './nav/NavItems.svelte';
+</script>
+
+<a href="/" class="c-logo">
+  <img src="/images/logo.svg" alt="Routify">
+</a>
+
+<nav class="c-navigation">
+  <NavItems />
+</nav>
